chore(ui): refactor rules API to typescript file
